trandf: periodic layer, gridbias fix (bounding box)

+++ IMLCV/test/test_CV_disc.py
@@ -1,16 +1,11 @@
 import os
 import shutil
-from cmath import nan
-from functools import partial
 
 import numpy as np
-from IMLCV.base.bias import Bias, BiasF, GridBias, NoneBias
 from IMLCV.base.CV import CV, CvFlow, Volume, dihedral, rotate_2d
 from IMLCV.base.CVDiscovery import CVDiscovery, TranformerUMAP
 from IMLCV.base.MdEngine import YaffEngine
-from IMLCV.base.metric import Metric, MetricUMAP, hyperTorus
-from IMLCV.base.Observable import Observable
-from IMLCV.base.rounds import RoundsMd
+from IMLCV.base.metric import Metric
 from IMLCV.launch.parsl_conf.config import config
 from IMLCV.scheme import Scheme
 from molmod import units
@@ -27,24 +22,38 @@ os.chdir(dname)
 
 def cleancopy(base):
 
+    if not os.path.exists(f"{base}"):
+        os.mkdir(base)
+        return
+
     if not os.path.exists(f"{base}_orig"):
         assert os.path.exists(f"{base}"), "folder not found"
-        shutil.copytree(f"{base}", f"{base}.orig")
+        shutil.copytree(f"{base}", f"{base}_orig")
 
     if os.path.exists(f"{base}"):
         shutil.rmtree(f"{base}")
     shutil.copytree(f"{base}_orig", f"{base}")
 
 
-def test_cv_discovery(name="test_cv_003", recalc=False):
+def test_cv_discovery(name="test_cv_disc_004", recalc=False):
     # make copy and restore orig
 
-    config(cluster='doduo', max_blocks=15)
+    config(cluster='doduo', max_blocks=14)
 
-    cleancopy(f'output/{name}')
+    full_name = f'output/{name}'
+    pe = os.path.exists(full_name)
 
-    if recalc:
+    components = 22.0
 
+    cvd = CVDiscovery(transformer=TranformerUMAP(
+        outdim=2, periodicity=[False, False], bounding_box=np.array([
+            [0.0, components],
+            [0.0, components],
+        ])))
+
+    if recalc or not pe:
+        if pe:
+            shutil.rmtree(full_name)
         T = 600*kelvin
 
         cv0 = CV(
@@ -58,54 +67,42 @@ def test_cv_discovery(name="test_cv_003", recalc=False):
                               [-np.pi, np.pi]])
         )
 
-        scheme0: Scheme = Scheme(cvd=CVDiscovery(transformer=TranformerUMAP()),
+        scheme0: Scheme = Scheme(cvd=cvd,
                                  cvs=cv0,
                                  Engine=YaffEngine,
                                  ener=get_alaninedipeptide_amber99ff,
                                  T=T,
                                  timestep=2.0 * units.femtosecond,
                                  timecon_thermo=100.0 * units.femtosecond,
-                                 folder=f'output/{name}',
-                                 write_step=20,
+                                 folder=full_name,
+                                 write_step=5,
                                  )
 
-        scheme0.round(rnds=2, steps=1e4, n=3)
-    else:
-        scheme0 = Scheme.from_rounds(
-            cvd=CVDiscovery(transformer=TranformerUMAP()),
-            folder=f'output/{name}',
-        )
+        scheme0.round(rnds=3, steps=5e3, n=5)
+
+        del scheme0  # close roundsobject
+
+    cleancopy(full_name)
+
+    scheme0 = Scheme.from_rounds(
+        cvd=cvd,
+        folder=full_name,
+    )
 
     scheme0.update_CV(
-        samples=2e3,
+        samples=5e3,
         # parametric_reconstruction=True,
-        global_correlation_loss_weight=0.6,
+        n_neighbors=50,
+        min_dist=0.9,
+        # densmap=True,
+        nunits=256,
+        nlayers=3,
+
+        # global_correlation_loss_weight=0.6,
         decoder=False,
+        # run_eagerly=True,
     )
-    scheme0.round(rnds=2, steps=3e4, n=4)
-
-    # base = f"output/{name}"
-
-    # cleancopy(base=base)
-
-    # rounds = RoundsMd.load(base)
-
-    # cvd = CVDiscovery()
-    # newCV = cvd.compute(rounds=rounds)
-
-    # # update the cv
-    # mde = rounds.get_engine()
-    # mde.bias = NoneBias(cvs=newCV)
-    # rounds.new_round(mde)
-
-    # do new umbrella
-    # scheme = Scheme.from_rounds(cvd=CVDiscovery( ), folder=f"output/{name}")
-
-    # scheme._grid_umbrella(steps=2e4, n=3)
-    # scheme._FESBias()
-
-    # scheme.rounds.new_round(scheme.md)
-    # scheme.rounds.save()
+    scheme0.round(rnds=3, steps=5e3, n=5)
 
 
 if __name__ == "__main__":
